+++ mapper.go
@@ -1,7 +1,9 @@
 package main
 
 import (
-	"github.com/iver-wharf/wharf-api-client-go/pkg/wharfapi"
+	"strconv"
+
+	"github.com/iver-wharf/wharf-api-client-go/v2/pkg/model/request"
 	"github.com/xanzy/go-gitlab"
 )
 
@@ -10,14 +12,14 @@ type mapper struct {
 	providerID uint
 }
 
-func (m *mapper) mapProjectToWharfEntity(proj gitlab.Project, buildDef string) wharfapi.Project {
+func (m *mapper) mapProjectToWharfEntity(proj gitlab.Project, buildDef string) request.Project {
 	groupName := ""
 
 	if proj.Namespace != nil {
 		groupName = proj.Namespace.FullPath
 	}
 
-	return wharfapi.Project{
+	return request.Project{
 		Name:            proj.Name,
 		BuildDefinition: buildDef,
 		Description:     proj.Description,
@@ -26,22 +28,13 @@ func (m *mapper) mapProjectToWharfEntity(proj gitlab.Project, buildDef string) w
 		TokenID:         m.tokenID,
 		ProviderID:      m.providerID,
 		GroupName:       groupName,
+		RemoteProjectID: strconv.Itoa(proj.ID),
 	}
 }
 
-func (m *mapper) mapBranchToWharfEntity(projID uint, branch gitlab.Branch) wharfapi.Branch {
-	return wharfapi.Branch{
-		ProjectID: projID,
-		Name:      branch.Name,
-		Default:   branch.Default,
-		TokenID:   m.tokenID,
+func (m *mapper) mapBranchToWharfEntity(branch gitlab.Branch) request.Branch {
+	return request.Branch{
+		Name:    branch.Name,
+		Default: branch.Default,
 	}
-}
-
-func (m *mapper) mapBranchesToWharfEntity(projectID uint, branches []*gitlab.Branch) []wharfapi.Branch {
-	var mappedBranches []wharfapi.Branch
-	for _, branch := range branches {
-		mappedBranches = append(mappedBranches, m.mapBranchToWharfEntity(projectID, *branch))
-	}
-	return mappedBranches
 }