+++ testdoubles/wharfclientapifetcher_mock.go
@@ -1,7 +1,9 @@
 package testdoubles
 
 import (
-	"github.com/iver-wharf/wharf-api-client-go/pkg/wharfapi"
+	"github.com/iver-wharf/wharf-api-client-go/v2/pkg/model/request"
+	"github.com/iver-wharf/wharf-api-client-go/v2/pkg/model/response"
+	"github.com/iver-wharf/wharf-api-client-go/v2/pkg/wharfapi"
 	"github.com/stretchr/testify/mock"
 )
 
@@ -11,55 +13,93 @@ type WharfClientAPIFetcherMock struct {
 	mock.Mock
 }
 
-// GetTokenByID returns a token as identified by its ID.
-func (m *WharfClientAPIFetcherMock) GetTokenByID(tokenID uint) (wharfapi.Token, error) {
-	args := m.Called(tokenID)
-	return args.Get(0).(wharfapi.Token), args.Error(1)
+// CreateProjectBranch adds a branch to the project with the matching
+// project ID by invoking the HTTP request:
+//  POST /api/project/{projectId}/branch
+func (m *WharfClientAPIFetcherMock) CreateProjectBranch(projectID uint, branch request.Branch) (response.Branch, error) {
+	args := m.Called(projectID, branch)
+	return args.Get(0).(response.Branch), args.Error(1)
 }
 
-// GetToken returns a token as identified by its token and username strings.
-func (m *WharfClientAPIFetcherMock) GetToken(token, userName string) (wharfapi.Token, error) {
-	args := m.Called(token, userName)
-	return args.Get(0).(wharfapi.Token), args.Error(1)
+// UpdateProjectBranchList resets the default branch and list of branches for a project
+// using the project ID from the first branch in the provided list by invoking
+// the HTTP request:
+//  PUT /api/project/{projectId}/branch
+func (m *WharfClientAPIFetcherMock) UpdateProjectBranchList(projectID uint, branches []request.Branch) ([]response.Branch, error) {
+	args := m.Called(projectID, branches)
+	return args.Get(0).([]response.Branch), args.Error(1)
 }
 
-// PostToken creates a new token and returns the created token,
-// populated with its newly assigned ID.
-func (m *WharfClientAPIFetcherMock) PostToken(token wharfapi.Token) (wharfapi.Token, error) {
-	args := m.Called(token)
-	return args.Get(0).(wharfapi.Token), args.Error(1)
-}
-
-// GetProviderByID returns a provider as identified by its ID.
-func (m *WharfClientAPIFetcherMock) GetProviderByID(providerID uint) (wharfapi.Provider, error) {
-	args := m.Called(providerID)
-	return args.Get(0).(wharfapi.Provider), args.Error(1)
-}
-
-// GetProvider returns a provider as identified by its name, URL, and upload
-// URL strings, as well as its token ID reference.
-func (m *WharfClientAPIFetcherMock) GetProvider(providerName, urlStr, uploadURLStr string, tokenID uint) (wharfapi.Provider, error) {
-	args := m.Called(providerName, urlStr, uploadURLStr, tokenID)
-	return args.Get(0).(wharfapi.Provider), args.Error(1)
-}
-
-// PostProvider creates a new provider and returns the created provider,
-// populated with its newly assigned ID.
-func (m *WharfClientAPIFetcherMock) PostProvider(provider wharfapi.Provider) (wharfapi.Provider, error) {
-	args := m.Called(provider)
-	return args.Get(0).(wharfapi.Provider), args.Error(1)
-}
-
-// PutProject creates or updates a project, based on wether the project ID value
-// is non-zero, and returns the created or updated project,
-// populated with its possibly newly assigned ID.
-func (m *WharfClientAPIFetcherMock) PutProject(project wharfapi.Project) (wharfapi.Project, error) {
+// CreateProject adds a new project to the database by invoking the
+// HTTP request:
+//  POST /api/project
+func (m *WharfClientAPIFetcherMock) CreateProject(project request.Project) (response.Project, error) {
 	args := m.Called(project)
-	return args.Get(0).(wharfapi.Project), args.Error(1)
+	return args.Get(0).(response.Project), args.Error(1)
 }
 
-// PutBranches replaces the list of branches for a project.
-func (m *WharfClientAPIFetcherMock) PutBranches(branches []wharfapi.Branch) ([]wharfapi.Branch, error) {
-	args := m.Called(branches)
-	return args.Get(0).([]wharfapi.Branch), args.Error(1)
+// GetProject fetches a project by ID by invoking the HTTP request:
+//  GET /api/project/{projectID}
+func (m *WharfClientAPIFetcherMock) GetProject(projectID uint) (response.Project, error) {
+	args := m.Called(projectID)
+	return args.Get(0).(response.Project), args.Error(1)
+}
+
+// GetProjectList filters projects based on the parameters by invoking the HTTP
+// request:
+//  GET /api/project
+func (m *WharfClientAPIFetcherMock) GetProjectList(params wharfapi.ProjectSearch) (response.PaginatedProjects, error) {
+	args := m.Called(params)
+	return args.Get(0).(response.PaginatedProjects), args.Error(1)
+}
+
+// UpdateProject updates a project by ID by invoking the HTTP request:
+//  PUT /api/project/{projectID}
+func (m *WharfClientAPIFetcherMock) UpdateProject(projectID uint, project request.ProjectUpdate) (response.Project, error) {
+	args := m.Called(projectID, project)
+	return args.Get(0).(response.Project), args.Error(1)
+}
+
+// GetProvider fetches a provider by ID by invoking the HTTP request:
+//  GET /api/provider/{providerID}
+func (m *WharfClientAPIFetcherMock) GetProvider(providerID uint) (response.Provider, error) {
+	args := m.Called(providerID)
+	return args.Get(0).(response.Provider), args.Error(1)
+}
+
+// GetProviderList filters providers based on the parameters by invoking the HTTP
+// request:
+//  GET /api/provider
+func (m *WharfClientAPIFetcherMock) GetProviderList(params wharfapi.ProviderSearch) (response.PaginatedProviders, error) {
+	args := m.Called(params)
+	return args.Get(0).(response.PaginatedProviders), args.Error(1)
+}
+
+// CreateProvider creates a new provider by invoking the HTTP request:
+//  POST /api/provider
+func (m *WharfClientAPIFetcherMock) CreateProvider(provider request.Provider) (response.Provider, error) {
+	args := m.Called(provider)
+	return args.Get(0).(response.Provider), args.Error(1)
+}
+
+// GetToken fetches a token by ID by invoking the HTTP request:
+//  GET /api/token/{tokenID}
+func (m *WharfClientAPIFetcherMock) GetToken(tokenID uint) (response.Token, error) {
+	args := m.Called(tokenID)
+	return args.Get(0).(response.Token), args.Error(1)
+}
+
+// GetTokenList filters tokens based on the parameters by invoking the HTTP
+// request:
+//  GET /api/token
+func (m *WharfClientAPIFetcherMock) GetTokenList(params wharfapi.TokenSearch) (response.PaginatedTokens, error) {
+	args := m.Called(params)
+	return args.Get(0).(response.PaginatedTokens), args.Error(1)
+}
+
+// CreateToken adds a new a token by invoking the HTTP request:
+//  POST /api/token
+func (m *WharfClientAPIFetcherMock) CreateToken(token request.Token) (response.Token, error) {
+	args := m.Called(token)
+	return args.Get(0).(response.Token), args.Error(1)
 }