+++ CHANGELOG.md
@@ -12,6 +12,18 @@ This project tries to follow [SemVer 2.0.0](https://semver.org/).
 	https://changelog.md/
 -->
 
+## v2.0.0 (WIP)
+
+- BREAKING: Removed support for `github.com/iver-wharf/wharf-api` v4.
+  Now requires a minimum of wharf-api v5.0.0. (#43)
+
+- Added support for `github.com/iver-wharf/wharf-api` v5.0.0. (#43)
+
+- Changed version of `github.com/iver-wharf/wharf-api-client-go`
+  from v1.3.1 -> v2.0.0. (#26, #43)
+
+- Added GitLab's internal project ID when adding project to database. (#43)
+
 ## v1.3.0 (2022-01-03)
 
 - Added support for the TZ environment variable (setting timezones ex.